getting-started: first-application: Updated doc to compile when copied
When I found the getting-started docs and started playing around with
capy, I noticed that the button example did not compile as it was
missing a few things. I added the window variable that was missing along
with the backend init and event loop that was referenced in the previous
examples. This should now compile with when copied and align with both
what the image expects and continue to build on the other examples.

Signed-off-by: Ashton Fritz <ashton.fritz@usask.ca>